Deletede unused files and foldes
