--- src/Chirp.CSVDBService/Chirp.CSVDBService.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <ItemGroup>
-    <ProjectReference Include="..\SimpleDB\SimpleDB.csproj" />
-    <EmbeddedResource Include="chirp_cli_db.csv" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SQLDB\SQLDB.csproj" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-</Project>
