--- src/Chirp.CSVDBService/chirp_cli_db.csv
@@ -1,13 +1,0 @@
-Author,Message,Timestamp
-ropf,"Hello/comma/ BDSA students!",1690891760
-rnie,"Welcome to the course!",1690978778
-rnie,"I hope you had a good summer.",1690979858
-ropf,"Cheeping cheeps on Chirp :)",1690981487
-lukas,***IntegrationTestStatement***,1695571669
-nielsfaber,Hello!!!,1695578274
-nielsfaber,Hello!!!,1695578900
-nielsfaber,Hello!!!,1695586550
-nielsfaber,Hej,1695586580
-Marius,Hello!!!,1695903715
-"Os, det er bare os","Den ene er vokset op med tv, den anden ude",467654767
-Marius,Hello!!!,1695908655
