--- src/Chirp/Readoptions.cs
@@ -1,7 +1,0 @@
-using CommandLine;
-
-namespace ReadOptionsNS 
-{
-    [Verb("read", HelpText = "Read all cheeps.")]
-    public class ReadOptions {}
-}
