--- src/SimpleDB/IDatabaseRepository.cs
@@ -1,8 +1,0 @@
-interface IDatabaseRepository<T> //Contains the base methods used later in CSVDatabase
-{
-    abstract static CSVDatabase<T> GetCSVDatabase();
-    void SetPath(string path);
-    void SaveToFile(T item);
-
-    IEnumerable<T> ReadFromFile();
-}
